+++ Sources/VComponents/Components/Item Pickers/Segmented Picker/VSegmentedPickerContent.swift
@@ -10,21 +10,34 @@ import SwiftUI
 // MARK: - V Segmented Picker Content
 @available(tvOS, unavailable)
 @available(watchOS, unavailable)
-enum VSegmentedPickerContent<Data, Content>
+enum VSegmentedPickerContent<SelectionValue, Content>
     where
-        Data: RandomAccessCollection,
-        Data.Index == Int,
+        SelectionValue: Hashable,
         Content: View
 {
     // MARK: Properties
-    case title(data: Data, title: (Data.Element) -> String)
-    case content(data: Data, content: (VSegmentedPickerRowInternalState, Data.Element) -> Content)
+    case title(data: [SelectionValue], title: (SelectionValue) -> String)
+    case content(data: [SelectionValue], content: (VSegmentedPickerRowInternalState, SelectionValue) -> Content)
     
-    // MARK: Properties
+    // MARK: Array API
     var count: Int {
         switch self {
         case .title(let data, _): return data.count
         case .content(let data, _): return data.count
+        }
+    }
+
+    var indices: Range<Int> {
+        switch self {
+        case .title(let data, _): return data.indices
+        case .content(let data, _): return data.indices
+        }
+    }
+
+    func firstIndex(of element: SelectionValue) -> Int {
+        switch self {
+        case .title(let data, _): return data.firstIndex(of: element)! // Force-unwrap
+        case .content(let data, _): return data.firstIndex(of: element)! // Force-unwrap
         }
     }
 }