+++ Sources/VComponents/Components/Item Pickers/Wheel Picker/VWheelPickerContent.swift
@@ -11,17 +11,16 @@ import SwiftUI
 @available(macOS, unavailable)
 @available(tvOS, unavailable)
 @available(watchOS, unavailable)
-enum VWheelPickerContent<Data, Content>
+enum VWheelPickerContent<SelectionValue, Content>
     where
-        Data: RandomAccessCollection,
-        Data.Index == Int,
+        SelectionValue: Hashable,
         Content: View
 {
     // MARK: Properties
-    case title(data: Data, title: (Data.Element) -> String)
-    case content(data: Data, content: (VWheelPickerInternalState, Data.Element) -> Content)
+    case title(data: [SelectionValue], title: (SelectionValue) -> String)
+    case content(data: [SelectionValue], content: (VWheelPickerInternalState, SelectionValue) -> Content)
     
-    // MARK: Properties
+    // MARK: Array API
     var count: Int {
         switch self {
         case .title(let data, _): return data.count